+++ src/gamelogic/rapidXML/RapidXMLAdapter.hpp
@@ -32,7 +32,7 @@ namespace GameLogic {
         int get_tile_width();
         int get_tile_count();
         vector<vector<int>> get_map();
-        vector<tuple<int, int, char*>> get_objects();
+        vector<tuple<size_t, size_t, char*>> get_objects();
         string get_entity_texture(int x, int y);
     private:
         vector<char> _buffer;
